Fix float conversion and double promotion warnings

+++ src/objects/Ship.cpp
@@ -18,15 +18,15 @@ void Ship::update(const State& state, double deltaTime)
 		direction -= deltaTime * 3.6;
 	}
 	if (state.keys[GLFW_KEY_SPACE]) {
-		position.x += cos(direction) * 20 * deltaTime;
-		position.y += sin(direction) * 20 * deltaTime;
+		position.x += static_cast<float>(cos(direction) * 20.0 * deltaTime);
+		position.y += static_cast<float>(sin(direction) * 20.0 * deltaTime);
 	}
 }
 
 math::Point2 randomPoint()
 {
-	math::Point2 p{static_cast<float>(rand()) / RAND_MAX * 100,
-	               static_cast<float>(rand()) / RAND_MAX * 100};
+	math::Point2 p{static_cast<float>(rand() / RAND_MAX) * 100,
+	               static_cast<float>(rand() / RAND_MAX) * 100};
 	LOG << p.x << ' ' << p.y << ' ' << rand() / RAND_MAX;
 	return p;
 }
